refactor: move enhanced log generator to /api/v2/logs
Restore /api/logs to original behavior so candidates already mid-assignment
are not affected. The richer dataset (variable volume, JSON and stack-trace
bodies, realistic attributes) now lives at /api/v2/logs. Update README,
otlp-logs-api README, and the index page to point at the v2 endpoint.
