+++ otlp-log-viewer/otlp-logs-api/src/app/page.tsx
@@ -1,7 +1,7 @@
 export default function Home() {
     return (
             <main>
-                <p>This app only exports an API endpoint accessible via <a href="/api/logs">{"/api/logs"}.</a></p>
+                <p>This app only exports an API endpoint accessible via <a href="/api/v2/logs">{"/api/v2/logs"}.</a></p>
             </main>
     )
 }
